requirements.txt
certifi==2018.4.16
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.0
newsapi-python==0.2.3
wordcloud == 1.5.0
pillow ==6.1.0
numpy==1.14.5
pandas==0.23.3
palettable==3.2.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
SQLAlchemy>=1.3.0
Werkzeug==0.14.1
gunicorn==19.7.1